requirements.txt
beautifulsoup4==4.4.0
Django==1.9.6
django-appconf==1.0.1
django-avatar==2.0
django-dbbackup==2.3.3
django-ical==1.4
django-mathfilters==0.1.3
django-resized==0.3.5
django-smart-selects==1.2.1
django-tinymce==2.2.0
django-crispy-forms==1.6.1
dropbox==6.2
html2text==2015.6.21
icalendar==3.9.2
mailchimp3==1.0.11
oauthlib==1.1.1
phonenumbers==6.3.0
Pillow==2.8.1
pygeoip==0.3.1
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
urllib3==1.14
tweepy==3.5.0
wsgiref==0.1.2
lxml===3.4.2
